services/training-controller/requirements.txt
# Core dependencies
sqlalchemy==2.0.25
asyncpg==0.29.0
sqlmodel==0.0.14
kubernetes==29.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Git operations for model promotion
GitPython>=3.1.40
PyYAML>=6.0

# HTTP client for ArgoCD API
httpx>=0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
